Move InjectsOnChannel
To sit with the rest of the injection stuff
